mult_invite/mult_invite/requirement.txt
alembic==1.14.1
attrs==22.2.0
autopep8==2.0.4
Babel==2.13.1
backcall==0.2.0
bcrypt==4.2.0
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.32.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==2.0.1
Faker==18.3.1
fastjsonschema==2.18.1
filelock==3.13.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
flask-paginate==2024.4.12
Flask-Principal==0.4.0
Flask-Session==0.8.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
fonttools==4.49.0
fqdn==1.5.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
graphviz==0.20.3
greenlet==3.0.3
gunicorn==19.7.1
h11==0.14.0
httplib2==0.22.0
humanize==4.10.0
hyperopt==0.2.7
idna==2.10
imageio==2.34.2
importlib-metadata==7.0.1
imutils==0.5.4
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
msgspec==0.18.6
mypy-extensions==1.0.0
mysqlclient==2.1.1
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
nltk==3.9.1
nodeenv==1.8.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.2
overrides==7.4.0
packaging==23.0
pandas==2.2.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.4.0
pip-tools==7.3.0
pipreqs==0.5.0
platformdirs==4.2.0
plotly==5.19.0
pluggy==1.0.0
prometheus-client==0.18.0
prompt-toolkit==3.0.40
protobuf==4.22.1
psutil==5.9.6
psycopg2==2.9.7
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pydeck==0.8.1b0
pygal==3.0.4
pygame==2.5.1
Pygments==2.16.1
PyJWT==2.6.0
PyMySQL==1.1.1
pyparsing==3.0.9
pypng==0.20220715.0
pyproject_hooks==1.0.0
pyright==1.1.344
pyrsistent==0.19.3
pytest==7.2.2
pytest-django==4.5.2
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.7
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-json-logger==2.0.7
python-socketio==5.11.4
pytz==2022.7.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzbar==0.1.9
pyzmq==25.1.1
qrcode==7.4.2
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
regex==2024.11.6
requests==2.24.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.12.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
simple-websocket==1.1.0
simplekml==1.3.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.4.3
stack-data==0.6.3
streamlit==1.32.0
streamlit-option-menu==0.3.12
stripe==11.2.0
syncthing==2.4.1
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.3.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.25.11
virtualenv==20.25.1
watchdog==4.0.0
wcwidth==0.2.9
webcam==1.34
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.1.3
whitenoise==6.4.0
widgetsnbextension==4.0.9
wsproto==1.2.0
WTForms==3.1.2
xgboost==2.0.3
yarg==0.1.9
zipp==3.17.0
